Refactor CartAnalytics
